--- src/ripple/app/tx/impl/details/PageUtils.h
@@ -1,45 +1,0 @@
-//------------------------------------------------------------------------------
-/*
-  This file is part of rippled: https://github.com/ripple/rippled
-  Copyright (c) 2021 Ripple Labs Inc.
-
-  Permission to use, copy, modify, and/or distribute this software for any
-  purpose  with  or without fee is hereby granted, provided that the above
-  copyright notice and this permission notice appear in all copies.
-
-  THE  SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
-  WITH  REGARD  TO  THIS  SOFTWARE  INCLUDING  ALL  IMPLIED  WARRANTIES  OF
-  MERCHANTABILITY  AND  FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
-  ANY  SPECIAL ,  DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
-  WHATSOEVER  RESULTING  FROM  LOSS  OF USE, DATA OR PROFITS, WHETHER IN AN
-  ACTION  OF  CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
-  OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
-*/
-//==============================================================================
-
-#ifndef RIPPLE_TX_IMPL_DETAILS_PAGEUTILS_H_INCLUDED
-#define RIPPLE_TX_IMPL_DETAILS_PAGEUTILS_H_INCLUDED
-
-#include <ripple/basics/base_uint.h>
-#include <ripple/basics/tagged_integer.h>
-#include <ripple/ledger/ApplyView.h>
-#include <ripple/protocol/AccountID.h>
-#include <ripple/protocol/TER.h>
-
-namespace ripple {
-
-/** Finds the token in the owner's token directory.  Returns token and page. */
-struct TokenAndPage
-{
-    STObject token;
-    std::shared_ptr<SLE> page;
-
-    TokenAndPage(STObject const& token_, std::shared_ptr<SLE> page_)
-        : token(token_), page(std::move(page_))
-    {
-    }
-};
-
-}  // namespace ripple
-
-#endif  // RIPPLE_TX_IMPL_DETAILS_PAGEUTILS_H_INCLUDED
